+++ src/main/java/net/lordofthetimes/characterCard/commands/RealNameCommand.java
@@ -1,6 +1,5 @@
 package net.lordofthetimes.characterCard.commands;
 
-import com.mojang.brigadier.suggestion.Suggestions;
 import dev.jorel.commandapi.CommandAPICommand;
 import dev.jorel.commandapi.arguments.ArgumentSuggestions;
 import dev.jorel.commandapi.arguments.GreedyStringArgument;
@@ -10,11 +9,9 @@ import net.lordofthetimes.characterCard.utils.MessageSender;
 
 public class RealNameCommand {
 
-    private final CharacterCard plugin;
     private final EssentialsXHook essentialsX;
 
     public RealNameCommand(CharacterCard plugin) {
-        this.plugin = plugin;
         essentialsX = plugin.essentials;
         new CommandAPICommand("realname")
                 .withAliases("rn","named")