+++ src/main/java/net/lordofthetimes/characterCard/hooks/CharacterCardPlaceholderExpansion.java
@@ -17,12 +17,10 @@ public class CharacterCardPlaceholderExpansion extends PlaceholderExpansion{
 
     private final CharacterCard plugin;
     private final DatabaseManager db;
-    private final YamlDocument config;
 
     public CharacterCardPlaceholderExpansion(CharacterCard plugin, DatabaseManager db, YamlDocument config){
         this.plugin = plugin;
         this.db = db;
-        this.config = config;
     }
 
     @Override