+++ src/main/java/net/lordofthetimes/characterCard/utils/ConfigMigration.java
@@ -1,11 +1,9 @@
 package net.lordofthetimes.characterCard.utils;
 
 import dev.dejvokep.boostedyaml.YamlDocument;
-import dev.dejvokep.boostedyaml.dvs.versioning.BasicVersioning;
 import dev.dejvokep.boostedyaml.settings.dumper.DumperSettings;
 import dev.dejvokep.boostedyaml.settings.general.GeneralSettings;
 import dev.dejvokep.boostedyaml.settings.loader.LoaderSettings;
-import dev.dejvokep.boostedyaml.settings.updater.UpdaterSettings;
 import net.lordofthetimes.characterCard.CharacterCard;
 
 import java.io.File;